EZP-26885: Moved Legacy StorageRegistry to more common namespace
